requirements.txt
fastparquet==2024.11.0
geopandas==1.0.1
pandas==2.2.3
psycopg2-binary==2.9.10
pyarrow==18.0.0
requests==2.32.3
